requirements.txt
astroid==2.4.2
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
colorama==0.4.3
fastapi==0.61.0
h11==0.9.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
motor==2.2.0
pycparser==2.20
pycryptodome==3.9.8
pydantic==1.6.1
pylint==2.5.3
pymongo==3.11.0
redis==3.5.3
six==1.15.0
starlette==0.13.6
toml==0.10.1
uvicorn==0.11.8
websockets==8.1
wrapt==1.12.1
